# Refactor filtering and optimize adaptive filtering