Also add a dependency on `multiversion` to take advantage of AVX2 for adaptive filtering when available.