@okaneco 2022-12-31
I was optimizing `sum_buffer` and I'm able to get better performance on `corpus-bench` only by changing that function than this PR. AVX2 might be a different story but the current computer I have access to doesn't support those instructions.

A comparison of the bench numbers for encoding on my computer. Make sure to `set_adaptive_filter` on the encoder.
```bash
$ cargo run --release --example corpus-bench -- ./tests/benches/
master         76 mps   0.09 GiB/s
this PR        84 mps   0.10 GiB/s
sum_buffer    148 mps   0.17 GiB/s
```

<details><summary>Optimized `sum_buffer` code</summary>

The `saturating_add` calls get optimized if you use an accumulator for the inner chunk loops. Not sure if this approach could help this PR.

```rust
fn sum_buffer(buf: &[u8]) -> u64 {
    const CHUNK_SIZE: usize = 32;

    let mut buf_chunks = buf.chunks_exact(CHUNK_SIZE);
    let mut sum = 0_u64;

    for chunk in &mut buf_chunks {
        // At most, `acc` can be `32 * (i8::MIN as u8) = 32 * 128 = 4096`.
        let mut acc = 0;
        for &b in chunk {
            acc += u64::from((b as i8).unsigned_abs());
        }
        sum = sum.saturating_add(acc);
    }

    let mut acc = 0;
    for &b in buf_chunks.remainder() {
        acc += u64::from((b as i8).unsigned_abs());
    }

    sum.saturating_add(acc)
}
```
</details>

Another thing I tried was keeping a running sum of each row's weights to return from `filter_internal` but doing that for `Avg` and `Paeth` degraded performance compared to summing the buffer after it was filtered. Keeping a running sum for `Sub` and `Up` was about the same or very slightly better performance.