@fintelia 2022-12-31
Interesting! I just tried your patch on my machine and also saw a clear benefit. Tweaking the `filter` method to skip considering the NoFilter case (which only serves to make compression worse!) and doing the filtering in-place both both give small additional gains in performance.
```
$ cargo run --example corpus-bench --release qoi_benchmark_suite
master                  83 mps   0.28 GiB/s
this PR (no AVX2)       86 mps   0.29 GiB/s
this PR (AVX2)         162 mps   0.55 GiB/s
sum_buffer             215 mps   0.73 GiB/s
sum_buffer + filter    232 mps   0.79 GiB/s
```

<details>
  <summary>Better filter() implementation</summary>

```rust
AdaptiveFilterType::Adaptive => {
    let mut min_sum: u64 = u64::MAX;
    let mut filter_choice = FilterType::NoFilter;
    for &filter in [Sub, Up, Avg, Paeth].iter() {
        filter_internal(filter, bpp, len, previous, current, output);
        let sum = sum_buffer(&output);
        if sum <= min_sum {
            min_sum = sum;
            filter_choice = filter;
        }
    }

    if filter_choice != Paeth {
        filter_internal(filter_choice, bpp, len, previous, current, output);
    }
    filter_choice
}
```
</details>