@okaneco 2022-12-31
I also see an increase with the improved `filter` :+1: